+++ src/domain/entity/PopupConfig.ts
@@ -1,6 +1,6 @@
-import type { PopupStyle } from "./PopupStyle.js";
 import type { EventType } from "../enum/EventType.js";
+import type { PopupStyle, PopupStyleOverride } from "./PopupStyle.js";
 
 export type PopupConfig = PopupStyle & {
-  events: Partial<Record<EventType, Partial<PopupStyle>>>;
+  events: Partial<Record<EventType, PopupStyleOverride>>;
 };